Stage 2/5: Let’s get metric Description The examples you've printed are great, but they are not really helpful. Let's take a step further and create a program that converts kilometers to meters. Remember that one kilometer equals exactly 1000 meters.
The program should prompt the user to enter the number of kilometers with the following message: "Enter a number and a measure: ". The user should input a number and a unit of measurement, for example, 2 km or 2 kilometers. Note that the units of measurement should be case insensitive, so "km", "KM", and "kM" should work just as fine as "km".

The program should then output the equivalent of x kilometers in meters in the following format: x kilometers is y meters.

If the name of the unit of measurement is entered incorrectly, the program should display: Wrong input.

Objectives
The program should convert the given number of kilometers to meters.
The user can input "km", "kilometer" or "kilometers".
The program should ask the user for input only once, not infinitely.
If the value is 1, the name of the unit must be singular, otherwise, plural: "1 kilometer" but "2 kilometers".
It is guaranteed that the number of kilometers is an integer.
If the name of the unit of measurement is entered incorrectly, the program should display: Wrong input.